chore : 클래스명 수정
